requirement.txt
absl-py (0.7.1)
astor (0.7.1)
attrs (19.1.0)
backcall (0.1.0)
backports.weakref (1.0rc1)
baidu-aip (2.2.17.0)
beautifulsoup4 (4.8.0)
bleach (1.5.0)
certifi (2019.11.28)
chardet (3.0.4)
colorama (0.4.1)
cycler (0.10.0)
Cython (0.29.7)
decorator (4.4.0)
defusedxml (0.6.0)
Django (2.0.10)
django-cors-headers (3.1.1)
djangorestframework (3.10.3)
dlib (19.7.0)
entrypoints (0.3)
et-xmlfile (1.0.1)
gast (0.2.2)
geoip2 (3.0.0)
google-pasta (0.1.7)
grpcio (1.20.1)
h5py (2.9.0)
html5lib (0.9999999)
idna (2.8)
ipykernel (5.1.1)
ipython (7.5.0)
ipython-genutils (0.2.0)
ipywidgets (7.4.2)
jdcal (1.4.1)
jedi (0.13.3)
Jinja2 (2.10.1)
jsonschema (3.0.1)
jupyter (1.0.0)
jupyter-client (5.2.4)
jupyter-console (6.0.0)
jupyter-core (4.4.0)
Keras (2.2.4)
Keras-Applications (1.0.7)
Keras-Preprocessing (1.0.9)
lxml (4.3.3)
Markdown (3.1)
MarkupSafe (1.1.1)
matplotlib (2.0.2)
maxminddb (1.5.2)
mistune (0.8.4)
mock (3.0.5)
mysqlclient (1.4.4)
nbconvert (5.5.0)
nbformat (4.4.0)
notebook (5.7.8)
numpy (1.18.1)
opencv-python (3.4.6+contrib)
openpyxl (3.0.0)
pandocfilters (1.4.2)
parso (0.4.0)
pickleshare (0.7.5)
Pillow (6.0.0)
pip (9.0.1)
prometheus-client (0.6.0)
prompt-toolkit (2.0.9)
protobuf (3.7.1)
Pygments (2.4.0)
pyparsing (2.2.0)
pypcap (1.2.1)
pyrsistent (0.15.2)
python-dateutil (2.6.1)
pytz (2017.2)
pywinpty (0.5.5)
PyYAML (5.1)
pyzmq (18.0.1)
qtconsole (4.4.4)
requests (2.22.0)
requests-file (1.4.3)
scapy (2.4.2)
scapy-http (1.8.2)
scipy (1.2.1)
Send2Trash (1.5.0)
setuptools (41.0.1)
six (1.11.0)
tensorboard (1.14.0)
tensorflow (1.14.0)
tensorflow-estimator (1.14.0)
tensorflow-hub (0.6.0)
termcolor (1.1.0)
terminado (0.8.2)
testpath (0.4.2)
tornado (6.0.2)
tqdm (4.32.0)
traitlets (4.3.2)
urllib3 (1.25.7)
wcwidth (0.1.7)
Werkzeug (0.15.2)
wheel (0.33.4)
widgetsnbextension (3.4.2)
wincertstore (0.2)
winpcapy (1.0.2)
wrapt (1.11.2)
